src/backend/requirements.txt

fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.2
email-validator==2.1.1
PyJWT==2.8.0
python-multipart==0.0.9
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
